web/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
